Fix CLANG errors introduced by 7d87f02799bd0a8fd36df24fab5baa4968615c86
Summary: Fix some CLANG errors introduced in 7d87f02799bd0a8fd36df24fab5baa4968615c86

Test Plan: Build with both of CLANG and gcc

Reviewers: rven, yhchiang, kradhakrishnan, anthony, IslamAbdelRahman, ngbronson

Subscribers: leveldb, dhruba

Differential Revision: https://reviews.facebook.net/D52329

+++ db/db_impl.cc
@@ -4387,7 +4387,8 @@ Status DBImpl::WriteImpl(const WriteOptions& write_options,
         pg.leader = &w;
         pg.last_writer = last_writer;
         pg.early_exit_allowed = !need_log_sync;
-        pg.running.store(write_batch_group.size(), std::memory_order_relaxed);
+        pg.running.store(static_cast<uint32_t>(write_batch_group.size()),
+                         std::memory_order_relaxed);
         write_thread_.LaunchParallelFollowers(&pg, current_sequence);
 
         ColumnFamilyMemTablesImpl column_family_memtables(
